+++ plugins/orchestrator-common/src/index.ts
@@ -6,3 +6,4 @@ export * from './models';
 export * from './workflow';
 export * from './QueryParams';
 export * from './utils/StringUtils';
+export * from './permissions';